# openapi.json
{
  "components": {
    "responses": {
...
      "InvalidToken": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The token is invalid or expired, or its user is inactive (`\"Invalid token\"`). The UI logs the user out on this response."
      },
...
      "RateLimited": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Too many requests from this IP (`\"Too many requests. Please try again later.\"`). `RateLimit-*` headers describe the window."
      },
      "ServerError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "An unexpected error (`\"Something went wrong. Please try again.\"`). Details are logged on the server, never returned."
      },
      "TokenNotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "No `Authorization` header was sent (`\"Token not found\"`)."
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "errors": {
            "properties": {
              "msg": {
                "example": "User not found",
                "type": "string"
              }
            },
            "required": [
              "msg"
            ],
            "type": "object"
          }
        },
        "required": [
          "errors"
        ],
        "type": "object"
      },
...
      "Media": {
        "properties": {
          "__v": {
            "type": "integer"
          },
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "album": {
            "type": "string"
          },
          "artist": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "author": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "cast": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "dateReleased": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "director": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "genre": {
            "type": "string"
          },
          "host": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "mediaId": {
            "description": "ID in the source catalog (TMDB, Spotify, Google Books).",
            "type": "string"
          },
          "mediaType": {
            "enum": [
              "MOVIE",
              "BOOK",
              "PODCAST",
              "TV",
              "MUSIC",
              "THEATRE"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "producer": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tagLine": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "name"
        ],
        "type": "object"
      },
      "MediaResponse": {
        "properties": {
          "data": {
            "properties": {
              "media": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "required": [
              "media"
            ],
            "type": "object"
          },
          "status": {
            "enum": [
              "success"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "data"
        ],
        "type": "object"
      },
...
      "ObjectId": {
        "example": "63019b905ccf53564ffb0c84",
        "pattern": "^[a-f0-9]{24}$",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "tokenAuth": {
        "description": "The JWT returned by signup/login, sent as the raw header value — no `Bearer ` prefix.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API for Rate It, a social app for rating movies, TV, music, and books. Success responses are `{ status: \"success\", data: { ... } }`; errors are `{ errors: { msg } }`.",
    "title": "Rate It API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/media/add": {
      "post": {
        "description": "Stores the media, or returns the existing item when one with the same `mediaId` and `mediaType` is already stored.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "media": {
                    "description": "Fields of the `Media` schema, without `_id`.",
                    "properties": {
                      "description": {
                        "type": "string"
                      },
                      "mediaId": {
                        "type": "string"
                      },
                      "mediaType": {
                        "enum": [
                          "MOVIE",
                          "BOOK",
                          "PODCAST",
                          "TV",
                          "MUSIC",
                          "THEATRE"
                        ],
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "picture": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "name"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "media"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "The media was already stored; returns the existing item."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Media created."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Add a media item",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/media/book/info/{googleBookId}": {
      "get": {
        "description": "Public. Returns the stored book, or fetches it from Google Books and stores it on first view.",
        "parameters": [
          {
            "in": "path",
            "name": "googleBookId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "The book was already stored."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Fetched from Google Books and stored."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Media not found\"`: Google Books returned no volume."
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Unable to fetch media details from Google Books\"`. Also returned for ids Google Books doesn't know, since it answers those with an error."
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Get book details",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/media/movie/info/{tmdbId}": {
      "get": {
        "description": "Public. Returns the stored movie, or fetches it from TMDB and stores it on first view.",
        "parameters": [
          {
            "in": "path",
            "name": "tmdbId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "The movie was already stored."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Fetched from TMDB and stored."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Media not found\"`: TMDB returned no movie."
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Unable to fetch media details from TMDB\"`. Also returned for ids TMDB doesn't know, since TMDB answers those with an error."
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Get movie details",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/media/music/info/{spotifyId}": {
      "get": {
        "description": "Public. Returns the stored track, or fetches it from Spotify and stores it on first view.",
        "parameters": [
          {
            "in": "path",
            "name": "spotifyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "The track was already stored."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Fetched from Spotify and stored."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Media not found\"`: Spotify returned no track."
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Unable to fetch media details from Spotify\"`. Also returned for ids Spotify doesn't know, since Spotify answers those with an error."
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Get song details",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/media/tv/info/{tmdbId}": {
      "get": {
        "description": "Public. Returns the stored show, or fetches it from TMDB and stores it on first view.",
        "parameters": [
          {
            "in": "path",
            "name": "tmdbId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "The show was already stored."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Fetched from TMDB and stored."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Media not found\"`: TMDB returned no show."
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Unable to fetch media details from TMDB\"`. Also returned for ids TMDB doesn't know, since TMDB answers those with an error."
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Get TV show details",
        "tags": [
          "Media"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}